added pheremones plus diminishing plus made graphics faster

+++ AntSim/Cell.cs
@@ -10,8 +10,8 @@ namespace AntSim
     {
         public Location Location { get; private set; }
         public int AvailableFood { get; set; }
-        public int FoodPheremone { get; set; }
-        public int HomePheremone { get; set; }
+        public float FoodPheremone { get; set; }
+        public float HomePheremone { get; set; }
         public bool IsHome { get; set; }
         public Ant Ant { get; set; }
